feat: summary screen
