+++ firebase/utils/formatDateNow.ts
@@ -1,0 +1,9 @@
+export const formatDateNow = () => {
+  const date = new Date();
+  return date.toLocaleDateString('pt-BR', {
+    weekday: 'long',
+    day: '2-digit',
+    month: '2-digit',
+    year: 'numeric',
+  });
+};
